requirements.txt
streamlit>=1.29.0
google-generativeai>=0.8.0
PyMuPDF>=1.23.14
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.4
pillow>=10.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=1.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
